Add my fork or omnibus-ruby to the Gemfile
I have added a feature to omnibus that allows for multiple omnibus
artifacts to share from a shared library_path. That library_path also
comes from another omnibus artifact...in this case it is from the
common python dependencies.

I have opened a PR which has yet to be accepted. Add my fork to this
Gemfile in the interim.