Removing cruft from Vagrantfile
For right now I will be concentrating on Ubuntu builds as those seem to
be the most widely supported. Therefore, for the sake of simplicity, I
have updated the Vagrantfile to remove things like freebsd and smartos
as well as all platforms other than 12.04 LTS. This will get augmented
in a future iteration - at this point this is still somewhat of a proof
of concept.